+++ Firestore/core/src/firebase/firestore/remote/stream.mm
@@ -19,7 +19,6 @@
 #include <utility>
 
 #include "Firestore/Source/Remote/FSTDatastore.h"
-#include "Firestore/core/src/firebase/firestore/remote/stream_operation.h"
 #include "Firestore/core/src/firebase/firestore/util/error_apple.h"
 #include "Firestore/core/src/firebase/firestore/util/hard_assert.h"
 #include "Firestore/core/src/firebase/firestore/util/log.h"
@@ -325,7 +324,7 @@ void WatchStream::UnwatchTargetId(FSTTargetID target_id) {
 std::shared_ptr<GrpcCall> WatchStream::DoCreateGrpcCall(
     Datastore* const datastore, const absl::string_view token) {
   return datastore->CreateGrpcCall(
-      token, "/google.firestore.v1beta1.Firestore/Listen");
+      token, "/google.firestore.v1beta1.Firestore/Listen", this);
 }
 
 void WatchStream::DoOnStreamStart() {
@@ -408,7 +407,7 @@ void WriteStream::WriteMutations(NSArray<FSTMutation*>* mutations) {
 std::shared_ptr<GrpcCall> WriteStream::DoCreateGrpcCall(
     Datastore* const datastore, const absl::string_view token) {
   return datastore->CreateGrpcCall(token,
-                                   "/google.firestore.v1beta1.Firestore/Write");
+                                   "/google.firestore.v1beta1.Firestore/Write", this);
 }
 
 void WriteStream::DoOnStreamStart() {
